feat(feedV2): generic deposit and withdraw types (#6189)
### Description

Introduce new generic types for the TX feed for deposits and
withdrawals.

`DEPOSIT` replaces `EARN_DEPOSIT` and `EARN_SWAP_DEPOSIT`.
`WITHDRAW` replaces `EARN_WITHDRAW`.

Why?

1. This allows directly supporting all the withdrawals and deposits for
all dapps that Zerion already knows about.
2. We can do this now without too much effort and keep a "simple"
backend implementation for the new `getWalletTransactions`

This is just for the new feed API with Zerion. 

### Test plan

- Tests pass

### Related issues

- Related to RET-1204
- Also see this Slack
[thread](https://valora-app.slack.com/archives/C029Z1QMD7B/p1729847488933829)
for context

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)