feat(feedV2): generic claim reward type (#6198)
### Description

Introduce new generic type for the TX feed for claim reward.

`CLAIM_REWARD` replaces `EARN_CLAIM_REWARD`

Why?

1. This allows directly supporting more claim reward TXs, not only for
Earn but also GoodDollar, etc.
2. We can do this now without too much effort and keep a "simple"
backend implementation for the new `getWalletTransactions`

Similar to what was done in
https://github.com/valora-inc/wallet/pull/6189

This is just for the new feed API with Zerion. 

### Test plan

- Tests pass

### Related issues

- Related to RET-1204
- Also see this Slack
[thread](https://valora-app.slack.com/archives/C029Z1QMD7B/p1729847488933829)
for context

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)